Kill -PRELEASE_MODE
Summary:
Ding dong! This annoying flag had to be manually supplied during a release or
would otherwise create botched artifacts that no-one could install and require
another patch release to fix it.

This now infers whether or not we are in a "release mode" based on the task
list. This is still slightly fragile in that there might be implicit ways of
invoking an upload task, but for now it's clearly a win.

Will update docs upon landing.

Reviewed By: IanChilds

Differential Revision: D6246145

fbshipit-source-id: cf8beb047d0c4e5816637104e848e6f36ee01dd0