Enable preallocate mount content in SectionsAdapter
Summary:
this makes it possible to pass a LayoutHandler for preallocating mount content to a SectionAdapter. The component trees inside the section hierarchy will use this handler for preallocation as before.
If no handler is passed to the SectionAdapter, there's no preallocation, which is the status quo. This diff doesn't change any functionality, it just makes the way for enabling preallocation when we want to.

Up next: provide a default layout handler for preallocation in the ComponentTree so that it's not necessary to pass one to enable it.

Reviewed By: jsendros

Differential Revision: D6220852

fbshipit-source-id: eb593803c1a2dd9e31d544fbe6453ac4581d48d2