Component state update when we build and measure the ComponentTree
Summary:
`OnCreateInitialState` is supposed to be called before `OnCreateTreeProp`. However, when a component is measured through `Component.measure()` by code, https://fburl.com/jgi2dzh4 , `component.applyStateUpdates()` will not be called so any `State` in `OnCreateTreeProp` will not be available.

This diff aims to address this problem by coupling the component state update with the `Layout.createAndMeasureTreeForComponent()` function.

Reviewed By: mihaelao

Differential Revision: D6283476

fbshipit-source-id: ab7ed994c4202a74c217511bf83ad4200665ca83