Allow height to be animated after initial mount
Summary:
Note: this is probably not the correct long term fix, but I'm landing this to unblock the bookmarks team.

There are cases with incremental mount + recyclerview where the LithoView gets mounted and should immediately animate: from the perspective of the component, the LithoView's height changed, but previously since the LithoView wasn't mounted, we don't observe a height change in measure. For now, I'm just removing the isExpectingHeightChange check, the correct long term solution might be to communicate from the component to the LithoView that the height is changing (e.g. get that off the LayoutState).

Reviewed By: marco-cova

Differential Revision: D6246903

fbshipit-source-id: d1d717915e357d8903498cff3efb9d104b17808d