Introduce InterStageStore cache
Summary:
This adds the concept of an InterStageStore that can be used to persist
annotation processor information across round and module boundaries. For now,
this is only used by the PropNameInterStageStore, which retains parameter
names as those are not available in Java 7 bytecode.

This requires quite a few modifications in various places so I added the
InterStageStore to make future additions of this sort easier.

The actual use of this API happens in follow-up diffs and is mainly to keep
commits atomic and make reviews easier.

Reviewed By: IanChilds

Differential Revision: D6312984

fbshipit-source-id: bae92da57376d4247036d40de61fa43a02f2cca8