Make InternalNode respect padding set by the user if background drawable doesn't have any padding (pre L devices)
Summary:
Right now if background drawable contains padding, this padding will be set to the InternalNode. The way InternalNode knows that drawable has padding - is by calling `Drawable#getPadding` which returns true if there is padding and false otherwise.
The problem is that prior to API 21, LayerDrawable always returns `true` regardless if there is padding or not (see https://goo.gl/gExcMQ).
This causes padding to be reset to `0` if you try to specify LayerDrawable on pre-L device.

Reviewed By: muraziz

Differential Revision: D6308948

fbshipit-source-id: 240eea842a72122d6e77263662247d57588e0995