Fix generator for child components
Summary:
I've been poking around in the builder generator and noticed that the check for
nested components is never statisfied as we're comparing a raw type
(`specModel.getComponentType() == Component`) with a parameterized type
(`Component != Component<?>`). This means that we never create a shallow copy
inside the builder, but assign the full component.

Reviewed By: IanChilds

Differential Revision: D6347474

fbshipit-source-id: 1c68499dbc64604dd2299ef72eab6baf5b0d649e