Fix cached prop application
Summary:
This took me a while to debug. I didn't realize that we collect the props as set
first to deduplicate. This means that I cannot rely on the final order to update
their names.

Instead, I defer the name updates till the very end now and do them in one big
stream.

Reviewed By: marco-cova

Differential Revision: D6350010

fbshipit-source-id: f54435e189ad5f424f5361e4bf0c6b5b355c260f