Add TestTextSpec
Summary:
I plan on adding test specs for all built-in widgets, but just use this one for
prototyping purposes for now. The specs won't actually compile for all of them
yet, anyway.

I'm not quite sure if this is the right call or not, but I'm putting them into
`c.f.l.widget`, instead of `c.f.l.testing.widget` in order to avoid having to
make the specs themselves public.

I wish I wouldn't have to do that, but the namespace pollution is actually
minimal as there's an inherent namespace through the `Test` prefix.

Reviewed By: IanChilds

Differential Revision: D6336685

fbshipit-source-id: 037e800cc5904d2c4e1da55d7c1554c1d4d1f442