Mount children too when locked for animations
Summary: When components are locked for animations, we make sure that we mount them regardless of whether they are visible or not. However, we do not do the same for their children, as when their children are laid out they have no way of being aware that they should be mounted. Therefore, we force them to be mounted by calling incremental mount using bounds the size of the entire view that is being mounted.

Reviewed By: astreet

Differential Revision: D6384960

fbshipit-source-id: 2b404391ca8672dc7a624f630ecd1e181eaa17ae