Add DI processor for testing
Summary:
Hello! Welcome to my stack. I hope the way this is broken up makes it easy to
review. The downside of smaller, contained commits is that it's sometimes
difficult to see the larger picture. So here's the idea:

This adds support for DI and parameterized Components for TestSpecs. In order to
properly test this, I first need a processor that can handle these cases. I
don't want this to rely on some massively complex internal processor, and
instead use one that I control.

This diffs adds a new processor just for testing purposes, that implements the
DI helper. It exists as stand-alone gradle module as we cannot depend on it
otherwise.

Reviewed By: IanChilds

Differential Revision: D6397884

fbshipit-source-id: 22cc31d0b41c63e87fced792d3072175572bda39