Exit Section StateUpdate Removal Function Early
Summary:
In a multithreaded environment, the instance variable `mPendingStateUpdates` would have been modified by other calls to `clonePendingStateUpdatesFromInstanceToLocal()` and `removeCompletedStateUpdatesFromInstance()`.

This means that the local variable `pendingStateUpdates` given to `removeCompletedStateUpdatesFromInstance()` will not always remain valid.

If we have a key from `pendingStateUpdates` that is no longer found in `mPendingStateUpdates`. We can be sure that `pendingStateUpdates` is no longer valid and we should exit the function.

Reviewed By: mihaelao

Differential Revision: D6346524

fbshipit-source-id: f7cfafd11a1821ad742cc41021b979d0d7e11f44