Provide better error messages for matchers
Summary:
This doesn't cover all cases, but the majority this should provide way more
useful error messages with test specs.

I'm adding `TextDescription`s which rely on the string representation of the
underlying matchers. I also chose to include the attribute that failed to match
against. This could potentially be confusing in cases where we scan an entire
tree and will highlight the last item rather than collecting results of all of
them, but it's a best-effort approach for now.

Reviewed By: IanChilds

Differential Revision: D6405703

fbshipit-source-id: 05d4163cfac236a5382f314f6a01ef5f10b51f0b