Update getRootContext to find underlying Application/Activity/Service
Summary: We'll be using this for pooling and we need to make sure that we are getting the right root context. The assumption was that each Activity/Service/Application has a different underlying ContextImpl, but for example in Robolectric that's not actually the case. I'm not convinced it's the case for all of android either.

Reviewed By: passy

Differential Revision: D6405669

fbshipit-source-id: 6f7a3c5e98438d5bea7476d1cba4adebef46b8ef