Allocate mount content from ComponentPools
Summary:
Context for this change: I want to track the number of allocations we make for each mount content in order to decide whether to preallocate (e.g. if we've already allocated 10 pieces of content for a pool of size 10, we probably don't want to preallocate anymore).

In order to do this, I want to allocate the content in the MountContentPool/ComponentPools, similar to what we now do for preallocation. In a follow up diff, I'll count total allocations when deciding whether to preallocate.

Reviewed By: IanChilds

Differential Revision: D6425641

fbshipit-source-id: a93793faf94029261fbbb6d88106b7a3760f8789