Extract interfaces for RecyclePool and MountContentPool
Summary: We want to expose the ability to use a custom pool implementation for your mount content. In order to do that, I want to extract an interface that someone can implement and back with whatever datastructure they want (this means not necessarily using RecyclePool). I also extracted a DebugInfo interface to make things work with Stetho.

Reviewed By: IanChilds

Differential Revision: D6436142

fbshipit-source-id: e1cdd6f1c5ca3d86b48475bfa613a1b93dde7ab9