Implement longpress for Spannable string in TextSpec.
Summary: Until now there was no way of detecting longclick in individual spans in Text component. Adding support for longclick in spannables passed to TextSpec. It works similar way as the View's onclick listener, i.e. we post Runnable with specific delay on touch down and that Runnable is fired as long as we are still touching that span area and didn't move out or released touch before the delay. On that runnable we deliver the `onLongClick()` callback that user specified.

Reviewed By: astreet

Differential Revision: D6447988

fbshipit-source-id: fe568dd106c76c1b894c41067c6e569e103ec26f