Add stetho deprecation message
Summary:
We are deprecating Stetho integration as it will allow us to focus on a new debugging
tool which better handles some of the patterns we see when people builds apps using Litho. We
are not ready to announce anything yet but hope that this allows us to get there more
quickly.

Reviewed By: passy

Differential Revision: D6447438

fbshipit-source-id: ba01681c0ce4f14ce91f89c9c589d87c5b59f9aa