Revert ComponentHost recycling
Summary:
There have been a number of tasks related to the recent addition of
ComponentHost recycling. After clarifying that this is not extremely urgent to
ship, we decided to revert this for now to mitigate the launch-blockers.

astreet will be back this Friday and we can revert this revert and apply
the necessary fixes on top of it.

The only changes made after `hg backout` were to adjust to recent refactors and
codemods, regarding the removal of `getLifecycle()` and removal of generics on
`Component`.

This reverts the stack of:

- D6436213	[litho] Add OnCreateMountContentPool to allow mount specs to define custom MountContentPools
- D6436142	[litho] Extract interfaces for RecyclePool and MountContentPool
- D6426850	[litho] Add configuration parameters for ComponentHost recycling
- D6425784	[litho] Base whether to preallocate on total allocations and not just pre-allocations
- D6425641	[litho] Allocate mount content from ComponentPools
- D6424121	[litho] Update ComponentHost pool size to 0
- D6408586	[litho] Context -> ComponentContext for ComponentPools.acquire/release
- D6408529	[litho] Pass lifecycle to acquireMountContent for consistency with release
- D6405858	[litho] Update mount content pre-allocation to only pre-allocate up to a limit
- D6405720	[litho] Refactor mount content pooling to prepare for preallocation
- D6405669	[litho] Update getRootContext to find underlying Application/Activity/Service
- D6397801	[litho] Extract getBaseContext to ContextUtils

Reviewed By: IanChilds

Differential Revision: D6496191

fbshipit-source-id: bcb17a56596ec7bbf49fd965b109241d5305a849