Rename ComponentLayout to ActualComponentLayout in some methods
Summary:
This is a step towards returning a Component rather than a ComponentLayout in onCreateLayout. In order to ease the migration, we'll want to change the meaning of ComponentLayout so that we don't have to change every single callsite. The "real" ComponentLayout will be temporarily renamed to `ActualComponentLayout` before being renamed back to `ComponentLayout` later on. `onMeasure` and `onBoundsDefined` will continue to use `ComponentLayout`, so we do their renaming here.

I won't land this quite yet - I'll try to land in a stack, or at least similarly timed to the change to onCreateLayout, since we don't want to be in this state for long.

Reviewed By: passy

Differential Revision: D6446559

fbshipit-source-id: 471b345cad6853c3e44a8c2fe54c5c23a22c9d1a