Enable DebugComponent for state testing
Summary:
State testing is currently incompatible with the sub-component extractor as it ensures that remaining tasks on the looper are run, which trigger release operations on `InternalNode`s.

This installs some escape hatches that can be used to avoid releasing and reinstalling layout roots.

This unlocks integrating state updates into a more fluid `assertThat` API.

Reviewed By: mihaelao

Differential Revision: D6520184

fbshipit-source-id: a8e3f3c987adc1a15c30b295fe66031acf3f34a0