Add withStateUpdate assertj helper
Summary:
Another one of these "Make the change easy, then make the easy change" ...
changes. This adds a fluent API for state updates. For now, it only covers the
simple, most used case, but this will be expanded in the future.

With Java 8, this reads a lot more nicely by the way:

```
assertThat(mComponentContext, mRecyclerCollectionComponent)
    .withStateUpdate(c -> RecyclerCollectionComponent.updateLoadingAndEmptyAsync(c, SUCCEEDED, true))
    .doesNotHave(
        deepSubComponentWith(
            anyOf(
                textEquals("loading"),
                textEquals("content"),
                textEquals("empty"),
                textEquals("error"))));

```

Reviewed By: IanChilds

Differential Revision: D6545230

fbshipit-source-id: 9a706f31cc8c4c35377cac2726f0ea107b83a95c