Save and restore raw props instead of processed props
Summary:
This fixes the issue when props are used in both delegate methods and event
methods.

The issue manifested itself before this stack when props were declared in multiple places, e.g. onMount and shouldUpdate. During a normal AP run, these would get extracted and then deduplicated based on their names. For test processor runs, the name resolution can fail. For that case, we have the prop name cache. However, that cache relies solely on the number and order of props. As the deduplication runs before the name application, we would apply the cached names on a wrongly deduplicated list, leading to all sorts of weird results.

This stack split the work up and saves a raw, unduplicated list first, that can be used to save and later restore the names of the props, irregardless of the availability of the names in a given run.

Reviewed By: IanChilds

Differential Revision: D6520077

fbshipit-source-id: fc4b7d7f6f4ccde725e020069e6375ea57d88b06