onCreateLayout returns a Component, not a ComponentLayout
Summary:
We're changing the API of LayoutSpecs, so that onCreateLayout returns a Component, not a ComponentLayout. This gives a nicer API (we always have a component, rather than sometimes having a ComponentLayout and sometimes a Component), and will lead to some nice optimizations in future around reconciliation.

In order to do this reasonably seamlessly:
1, ComponentLayout is now an interface that is extended by Component. This means that we don't have to change every onCreateLayout call right now (coming in a future diff).
2, What was ComponentLayout is renamed to ActualComponentLayout.
3, Row, Column, Wrapper, RowReverse and ColumnReverse become Components.
4, We have a resolve method that is used to resolve a Component's onCreateLayout method to a tree of InternalNodes.

Reviewed By: mihaelao

Differential Revision: D6447032

fbshipit-source-id: 10285977a34484d527f3fc100b87b8687082b4aa