Add ComponentScript specific solution for recursive onCreateLayout
Summary: previous, onCreateLayout was entirely recursive, but the changes in D6447032 mean that is no longer the case. This is the simplest solution I have to make CS scopes still work fine. I'll add the changes that are required to make the logging work too, but wanted to run the tests to see if this fixes the scope problem.

Reviewed By: michalgr

Differential Revision: D6537844

fbshipit-source-id: 0b063d8f0a39cc38da54c9fc4ad6deb2ed6ee2b8