Protect against recycling experiment for ComponentHost getting turned on while app is running
Summary: If the user gets a config update that turns on the recycling experiment while there are ComponentHosts that have been created but not recycled, we could try to release a ComponentHost that has a parent. The hack for now is to just ignore ComponentHosts that have a parent when release is called, which should be safe since we'll just do what we were doing before this experiment.

Reviewed By: IanChilds

Differential Revision: D6557385

fbshipit-source-id: 5230ae5299b65ba8af480dc852c46f91376a2dbb