Make EditText and backgrounds with padding work
Summary: EditTextSpec.onMeasure uses a fresh EditText for each measure. EditText has a default background which has padding, and if we also set a background on the layout, then both lots of padding get applied. Therefore, if EditTextSpec has a background that has padding, then we want to the measure the EditText without any padding.

Reviewed By: marco-cova

Differential Revision: D6396406

fbshipit-source-id: 3144777892bb4c41dbc4c3f8c50d1d57b3c07f7e