Rename updateState -> updateStateSync in ComponentTree/ComponentContext/SectionContext
Summary: This is the first step in our work to try to change the default updateState mechanism to be async instead of sync. After this, we'll add an alias to updateState named updateStateSync and post about the differences and when to use one or another. Finally, we'll codemod all existing updateState usage to updateStateSync and temporarily remove updateState (??... not sure what we actually want to do here after the refactor to keep people from adding more updateState calls).

Reviewed By: IanChilds

Differential Revision: D6569023

fbshipit-source-id: 620190bc52cb2c157d79b271cc2a966c05167feb