add representation of internal state update policy that could be supplied to EditText component
Summary:
Added EditTextStateUpdatePolicy class that takes care of creating a spec that could be supplied to EditText component, specifying when and how the later should update its internal state. This does violate encapsulation, but is essential for optimization, so costly state updates, which trigger relayout, happen only when is really needed.

Motivation:
For example, we need to create a EditText components that has a static width, but should dynamically expand in height as user types in (WRAP_CONTENT effect). In this case we should force relayout when number of lines changes. Also to ensure that no input changes aren't lost when something else triggers relayout of the component tree, like hiding a soft keyboard (this is done using lazy updates)

Reviewed By: mihaelao

Differential Revision: D6570650

fbshipit-source-id: 6423d7780fb3705c2841af608181fce4dcfca1fb