Propagate fetch location to BaseGraphQLConnectionSection
Summary:
We want to propagate the location of the fetch up to the `SectionTree`. A new `LoadingEvent.LoadingState` is created to ensure that we don't break existing contracts.

This will be a requirement for D6566314, where we try to batch focus/scroll requests when we need to wait for new data coming in from refresh.

Differential Revision: D6565973

fbshipit-source-id: 61169921f549bd9c8d0b81e219ca71ff5678c5d7