Focus Request Dispatcher
Summary:
Users will attempt to make a focus request to a section, the right Section's callback to do it is in `onDataBound()`. However, calling a focus request from outside of the Section hierarchy does not allow the user to do it under the data has been bounded.

This diff will allow all the focus requests to happen immediately, or the last one was saved to be processed after the data is bounded. This removes the need for the users to post a runnable on the UI handler to see their focus request works.

This feature will be used in D6566314

Reviewed By: mihaelao

Differential Revision: D6566314

fbshipit-source-id: e06d569bc7dedfc6772343e87b6b48e9714a6e54