Enable setting tooltips on nested anchor components
Summary:
currenly we are only searching for the anchor component by looking at the keys of direct children of the component that is the root of the layout. With Row/Column being treated as components that can also have keys, we need to handle searching through children that can be on deeper levels of the layout because this case will occur much more often.

This does a best-effort search of the anchor component by looking through children components of the root of the layout and anchoring on the first component that we find in a breadth-first search that matches the key with the anchor key.

Reviewed By: IanChilds

Differential Revision: D6602102

fbshipit-source-id: ca403f1aae2d4fe84ebee25f7b35faed8f2a5c82