Use scoped context in onMeasure
Summary: The context that we use on an InternalNode isn't necessarily the same as the one that is set to the Component. If we use the wrong context then the scopedComponent on that context will be wrong, and therefore events will not work. Instead, if we use the scoped context from the component then we will always use the correct context.

Reviewed By: mihaelao

Differential Revision: D6649551

fbshipit-source-id: 6a8ee9cc9fac0f60bdf7127c91fcfce7906303e0