Use synchronized pool for the map data structure in SectionsPools
Summary:
We are getting ` java.lang.IllegalStateException: Already in the pool!` error from `SectionsPools`.

I was not able to replicate this error. However, looking at the error [[ https://our.intern.facebook.com/intern/sapienz/crash/398320 | log ]] from Sapienz suggests that there is concurrency issue with not using `SynchronizedPool` in `RecyclePool`.

Sidenote: the reason why I did not use `SynchronizedPool` in the first place is because we are only accessing `SectionsPools.acquireStateUpdatesMap()` and `SectionsPools.release()` in a `synchronized` block within `SectionTree`.

Reviewed By: IanChilds

Differential Revision: D6665218

fbshipit-source-id: 92616e65e2d1ccd7afc2f17d1ae08d1d66f56290