When testing for running disappear animations, just use pending animation count
Summary:
D6650165 shows a bug where after a disappear animation, a view is still clickable. This is because there's a bug where we never complete a set of disappear animations because at the time an animation's finish listener is called, its isActive method still returns true.

Looking at the logic, it seems safe to just use the pending animation count as that factors in all running animations already.

Reviewed By: sergeynv

Differential Revision: D6692237

fbshipit-source-id: 1a73080ec9719862e1de122e7cbc2c60bede8d8d