revert scroll last scroll position fix
Summary:
Revert D6651210. We will find another approach for it.
The diff above will cause an issue that every time when onBoundsDefinied is triggered, the scroll position will be reset to 0. An example is when you go to search video tab, scroll the pillter hscroll bar, and then scroll down the search results below, the pillter bar will be reset. (T25058683)

Differential Revision: D6705143

fbshipit-source-id: d40bfdbdc7d50f24164d1bcb425b6759209a8697