Only apply copyProps one time, not twice
Summary: There was a (perf) bug where we were copying the same props to the same node twice, which this diff fixes. Previously, if we went through the `isInternalComponent` branch we would apply the common props in the resolve method, and also later on in createLayout. Now, we only apply the props in resolve if we're not then going to apply them later on too.

Reviewed By: astreet

Differential Revision: D6722369

fbshipit-source-id: f7de272d5f1ce5c364ee6ef222b07fec9b1ba797