Set up onError delegation for onCreateLayout(WithSizeSpec)
Summary: This delegates general exceptions caught in onCreateLayout and onCreateLayoutWithSizeSpec to the `onError` callback. This is guarded by a config flag which can be controlled through an experiment. This high-level switch is to analyze the potential performance impact as the implications of broad try/catch blocks in different JVMs are poorly documented.

Reviewed By: IanChilds

Differential Revision: D6738771

fbshipit-source-id: 8e28bdc801ea053a1f6225588fee18717d535359