Add option to trim tail and head for DataDiffSection
Summary:
This add an option to trim the items that pass a comparison check from the head and tail of the previous and next data of a DataDiffSection. Based on  pasqualeanatriello's D6231542
The comparison in this diff is a no-op, this just sets up the trimming logic. I'll add the comparison logic in the next diff.

Reviewed By: passy

Differential Revision: D6736759

fbshipit-source-id: 03dcb4431e3435c6b4dbb5a11c704da10693d061