Add LazyComponentSelector
Summary:
This diff adds a utility class to reduce boilerplate code around making lots of Component.willRender calls wrapped in if else statements.

This Component is equivalent to ComponentSelector but takes ComponentCreators as arguments instead of Components to avoid the unnecessary instantiation of Components that to do not get selected.

The test code shows sample usage.

Differential Revision: D6760356

fbshipit-source-id: b81ec57546f84ec38021cc70e0e427d39fa02732