InputFilter Props for EditTextSpec
Summary:
We will allow the users to specify their own `InputFilter` as an optional prop to `EditTextSpec`. As the prop is a vararg argument type, the users will be able to specify more than 1 `InputFilter`.

We also ensure that a `InputFilter.LengthFilter()` of length `maxLength`, which is also a prop, will be given to the underlying `EditText`

Related group post: https://fb.facebook.com/groups/1430200360634661/permalink/1981856755469016/

Reviewed By: passy

Differential Revision: D6736589

fbshipit-source-id: 9736be166f561c6b4b322291154a4ab50cd3a17c