Trigger Self Function
Summary:
Based on the post: https://fb.facebook.com/groups/1430200360634661/permalink/1975071909480834/

We want to be able to use the `Trigger` APIs split the number of `Props` and `State` required in `onCreateChildren()` of `Section`.

Before we can have it in `Section`, we first have to bring it onto `Component`.

This diff will allow the component to call its own `Trigger` function

Example:

```
public class TestLayoutSpec {

  OnCreateLayout
  static Component onCreateLayout(
      ComponentContext c) {
    return Column.create(c)
      .child(TestLayout.generateComponentA(c))
      .child(TestLayout.generateComponentB(c))
      .build();
  }

  OnTrigger
  static Component generateComponentA(
      ComponentContext c,
     Prop PropA propA,
     PropB propB,
     InjectProp ComponentA componentA) {
   return componentA.create(c).propA(propA).propB(propB).build();
  }

  OnTrigger
  static Component generateComponentB(
      ComponentContext c,
     Prop PropC propC,
     PropD propD,
     InjectProp ComponentB componentB) {
   return componentB.create(c).propC(propC).propD(propD).build();
  }
}
```

Reviewed By: mihaelao

Differential Revision: D6772727

fbshipit-source-id: 6ae3f28d11a51262bbda9de1c7c5e24b8c1fc702