Fix newEventHandler override
Summary:
This is a bit weird of a weird one. `newEventHandler` exists in both
ComponentContext and SectionContext. It looks like the latter is overriding the
former but isn't really. This even causes a warning to show up in logcat:

```
method Lcom/facebook/litho/sections/SectionContext;.newEventHandler incorrectly overrides package-private method with same name in Lcom/facebook/litho/ComponentContext;
```

This makes sure that we're at least explicit about it.

It is only used in the corresponding Section or Component Lifecycle classes, so
there should be no behavioral change.

Reviewed By: mihaelao

Differential Revision: D6807319

fbshipit-source-id: 334db2f40170fb4cb890982e4c68ac4f948c7b0d