Add a new Common Prop for Accessibility Role
Summary: The Android Accessibility API has a hidden concept of "Roles", which are determines solely by the class of a view.  These roles change the behavior of accessibility services, such as adding the string "Button" and "Double tap to activate" on elements with a Button role.  This is something that was only possible to set on Litho components by understanding this system and overriding the onInitializeAccessibilityNodeInfo handler before.  This diff makes accessibilityRole a common prop with a simple enum of values.  This role alone should not indicate that an element should be "accessible", but if an element is accessible already, the role should take effect.

Reviewed By: IanChilds

Differential Revision: D6816303

fbshipit-source-id: 2c2bea07731954fa3ad61f897de3ec901260e2b2