Make sure contentDescription alone is enough to trigger focusability
Summary: Currently contentDescription alone is enough to trigger Litho to create a wrapper view around the component for accessibility.  Unfortunately, this wrapper view was not always helpful, because it was not inherently focusable.  This changes the behavior so that any view that has a contentDescription, and is not explicitly set to not be important for accessibility, will be set to important, and therefore focusable.  This matches the behavior in the native Android View class (https://github.com/aosp-mirror/platform_frameworks_base/blob/master/core/java/android/view/View.java#L8604).

Reviewed By: IanChilds

Differential Revision: D6799219

fbshipit-source-id: d3570f4a32d2e30789009fde26bbec023fc05917