The root of a ComponentTree should be created non-null
Summary:
We're hitting crashes where `mRoot` is null when we attach the LithoView that the ComponentTree is set on.

There are only three methods in `ComponentTree` that mutate `mRoot` - the constructor, `setComponent` and `release()`. The logging I added previously verifies that this is not being caused by `release()`, as the `ComponentTree` has not been released. `setComponent` is also not responsible as it will only set `mRoot = root` is `root` is not null. Therefore I believe that it must be the case that someone is creating a `ComponentTree` with a null root.

Reviewed By: marco-cova

Differential Revision: D6807362

fbshipit-source-id: ad0eb9d1ff2f994af41da1f9e2ea0cb884a538b8