Fixed call to setImportantForAccessibility on API 15
Summary: the setImportantForAccessibility method was added in API 16, which means this code would crash any device on API 15.  Surprisingly, no one had noticed this yet, as this method had no call sites until D6799219.  Since we are using the ViewCompat library for all of the other importantForAccessibility calls, the simple fix is just to convert this one as well.

Differential Revision: D6836071

fbshipit-source-id: 3626a9b97310a91851ea170c5ed9c1d4777e50c2