Use StringDef for AccessibilityRole
Summary:
Since we cannot use an enum for the definition, we can add at least the illusion
of type safety by using a StringDef. This means we cannot use non-constant
expressions, but since we already had one exception for ViewPager, we might as
well apply this to all class names. In the unlikely event of one of these class
names changing, we would likely run a large `s/android\.widget\..../.../g`
anyway.

I also removed the `null` as it's a) non-constant and b) not useful to alias as
it should indicate the *absence* of a value which `NONE` would not be.

Reviewed By: IanChilds

Differential Revision: D6844567

fbshipit-source-id: b3c956f1d0a331d015aefeb9a29c6c59ced1ed9d