Fix type erasure in Components AP
Summary:
The old version of this code was slightly unsafe.  It was attempting to ensure
that expectedReturnType was a subclass of T doing a normal cast to T after the
runtime cast to expectedReturnType, but since T is type-erased, this second
cast actually did nothing.  The proper way to handle this is to declare up
front that expectedReturnType must extend T, which will allow javac to verify
this at the call-site.  Then the second cast can just be removed.

Differential Revision: D6849794

fbshipit-source-id: 47609f61be57125472950ad9f3465a62535d182b