@TreeProp and @InjectProp Declaration in Trigger Methods
Summary:
Previously, parameters annotated with `TreeProp` and `InjectProp` that were only used in `OnTrigger` methods were not being handled properly in the generated `Component` code.

This was because the generator was not taking `Trigger` methods into consideration when it was generating `TreeProp` and `InjectProp` code for the `Component`.

This diff will allow those annotated parameters to work correctly in Trigger.

Differential Revision: D6807189

fbshipit-source-id: ebd46104c56a98d3d940f89401ee789fc3d9c966