Added CommonProp for 'selected' state
Summary: We currently have common props for enabled and focused, but not one for selected.  This state is useful for accessibility services, as it can indicate a number of things, such as the state of tabs (which one you are on), the state of toggled buttons, the state of items in a list, etc.

Reviewed By: IanChilds

Differential Revision: D6789921

fbshipit-source-id: 97961e7fd640cdaf6ef23b95730bf1163acec767