Add delay to state update layout calculation
Summary:
We want to test the impact of changing the duration for the layout calculations triggered by state updates.
This is part of assessing the impact of reconciliation and only recalculating the part of subtree that triggered a state update.
We'll measure the layout calculation for a tree and delay the state updates with a percentage of that duration.

Reviewed By: pasqualeanatriello

Differential Revision: D6871848

fbshipit-source-id: 0cdfbef2d01d057a5d4c5d34bc6da6705e3afc4f