Experiment with unmounting all content of LithoView when it is recycled by RecyclerView
Summary:
Currently we unmount the contents of LithoView only through incrementalmount mechanism. That is, in every scroll we check every LithoView in RecyclerView and see overlap of RecyclerView's viewport and LithoView's bounds. Whatever part of LithoView that doesn't overlap with RecyclerView's viewport we unmount that part.

However, the last component cannot get unmounted by that mechanism because once it is recycled we cannot apply incrementalmount to it. When we are scrolling up the topmost component of last LithoView in the RecyclerView's list will not get unmounted. The same is true for the bottommost component of first LithoView when we scroll on the opposite direction.

While it sounds reasonable to unmount all items when view gets recycled, we need to experiment with it just to make sure it doesn't break anything.

Reviewed By: pasqualeanatriello

Differential Revision: D6875207

fbshipit-source-id: 71547e4b506f6d3e8f02cecd01a6743492b93fb7