Wait for RecyclerView updates before calculating range
Summary:
it turns out that sometimes RecyclerView also posts operations to next frame, which means that even if we post range layout calculation to next frame, async layout might still happen for items on screen because the compute range runnable gets posted before the RecyclerView sync operations.
This accounts for that and waits until the RecyclerView has no more pending operations before calculating the range on the bg thread.

Reviewed By: pasqualeanatriello

Differential Revision: D6923636

fbshipit-source-id: 2b4561e25c2d19df771364def3c8a92bf96fae29