Add ability to make the talkback description of a TextSpec different from the text within it
Summary:
Originally the TextSpec's talkback description was intuitively set to be the same as the text contained. In this diff I'm adding support for making the content description different than the text if you pass in a separate prop for the talkback.

The reasoning for this is because in the case of Lists, two things can be removed - a list item can be removed and the entire list attachment can be removed. We have "X" buttons to remove individual list items and we have a button that says "Remove List" that is used to remove the entire list attachment. In talking to our team's content strategist, people tend to move through screen readers/talkbacks quickly and we therefore want the talkback for removing an entire list to be "Remove whole list" to make it immediately distinguishable from "Remove list item".

Reviewed By: blavalla

Differential Revision: D6930374

fbshipit-source-id: 2b5b606c08d90146f2f16deae8e9e64e18adc17c