Make sure ViewBinder.unbind is called on recycle
Summary:
Not sure if setting a ViewBinder on the BaseViewHolder is ideal, but it's probably the safest thing to do.

I considered doing
```java
mComponentTreeHolders.get(holder.getAdapterPosition()).getRenderInfo()
    .getViewBinder().unbind(viewHolder.itemView)
```

but that led to some crashes likely due to synchronization issues where mComponentTreeHolders' contents could update before the UI would update and we'd lose the reference to the correct ViewBinder.

Fixes #305.
Closes https://github.com/facebook/litho/pull/308

Reviewed By: passy

Differential Revision: D6936402

Pulled By: muraziz

fbshipit-source-id: 13f7f1f4b277dbd51666d05f18a534e49b803585