Check for Null EventTriggerTarget Instance
Summary:
There are cases where the `Trigger` generated can be used before the first Component layout is completed. This will result in a null exception.

We should exit gracefully when the this happens. We do not have to be worried about missing triggered events here because `ComponentContext` will be invalid in `OnTrigger` functions anyway.

Differential Revision: D6962077

fbshipit-source-id: 12626a82e52133957113d1324ad31ca59110884a