Fix StartSnapHelper when child views are smaller than half the size of the RecyclerView
Summary: We were chosing the view to snap to position for based on the closest to the center instead of the first one before the start of the RecyclerView. This made it so after a snap, we'd have a new view closest to the center, causing us to snap again. This would loop until we got to the end of the list. Instead, we now base the snap delta off the position of the first view whose start is before the recyclerview start.

Reviewed By: mihaelao

Differential Revision: D6988153

fbshipit-source-id: d93f1694d10f8a2761c481540d35bc62a426e4fb